chore: 폴더 정리
- header 컴포넌트 widgets 폴더에 추가
- app/ui -> app/product/ui 이동
- useFetchProductQuery product/hook 폴더로 이동
