+++ src/components/index.ts
@@ -1,3 +1,3 @@
 export { default as Card } from './Card';
 export { default as Button } from './Button';
-export { default as Header } from './Header';
+export { default as Badge } from './Badge';
