+++ src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { default as useFetchProductsQuery } from './useFetchProductsQuery';
 export { default as useIntersectionObserver } from './useIntersectionObserver';
